\name{lsr-package}
\alias{lsr-package}
\alias{lsr}
\docType{package}

\title{
Companion Package to "Learning Statistics with R" 
}

\description{
Tools for novice users, including wrapper functions for several statistical tests, simple effect size calculations, and basic data manipulation
}

\details{
\tabular{ll}{
Package: \tab lsr\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2015-02-18\cr
License: \tab GPL-3 \cr
LazyLoad: \tab yes\cr
}

The lsr package is a companion to the lecture notes from which I teach my introductory statistics class. It takes its name from the name of the lecture notes that I teach from, "Learning Statistics with R". The purpose of the package is to provide a few tools that are useful for students new to both statistics and to R. 

WARNING: This package is still under development. The functions work for most situations that I envisage them being used for, but some extensions are planned. In particular, for the moment there are no guarantees of backwards compatibility. Until this notice disappears some care is advised.

}

\author{
Author: Daniel Navarro
\cr
Maintainer: Daniel Navarro <daniel.navarro@adelaide.edu.au>
}

\references{
Navarro, D. J. (2015). Learning statistics with R: A tutorial for psychology students and other beginners. (Version 0.5). University of Adelaide. Adelaide, Australia. \url{http://ua.edu.au/ccs/teaching/lsr}

}

\keyword{ package }

%\seealso{
%}

%\examples{
%## ~~ simple examples of the most important functions ~~
%}
