\name{lsr-package}
\alias{lsr-package}
\alias{lsr}
\docType{package}

\title{
Companion package to "Learning Statistics with R" 
}

\description{
Functions that I have found useful for teaching purposes.
}

\details{
\tabular{ll}{
Package: \tab lsr\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2013-09-04\cr
License: \tab GPL-3 \cr
LazyLoad: \tab yes\cr
}

The lsr package is a companion to the lecture notes from which I teach my introductory statistics class. It takes its name from the name of the lecture notes that I teach from, "Learning Statistics with R: A tutorial for psychology students and other beginners". The purpose of the package is to provide a few functions that are useful for beginners, but not necessarily needed by experienced users.

--- WARNING: This package is still under development. The functions work for most situations that I envisage them being used for, but some extensions are planned. Additionally, for the moment there are no guarantees of backwards compatibility. For instance, until this warning notice disappears, if I decide that a function is unnecessary, I will delete it rather than deprecate it. Finally, note that very little exception handling has been written at this stage. In short, although the package serves its basic goals (it works for all examples I need for my class), until this notice disappears some care is advised in the use of this package. ---

}

\author{
Author: Daniel Navarro
\cr
Maintainer: Daniel Navarro <daniel.navarro@adelaide.edu.au>
}

%\references{}

\keyword{ package }

%\seealso{
%}

%\examples{
%## ~~ simple examples of the most important functions ~~
%}
